src/LearnLanguage.AI/requirements.txt
openai
flask
flask-cors
python-dotenv
huggingface_hub
transformers
torch
scipy
faster-whisper
numpy
requests